Fix MSVC warning in PD0.C and optimize CI workflow
PD0.C: Cast is.get() to uint8_t to fix C4244 warning

CI optimization:
- Reduce Linux compiler tests from 10 to 4 (GCC 9+14, Clang 16+19)
- Use macos-latest and macos-26 (beta) for macOS testing
- Keep container builds (AlmaLinux 8, Rocky 9) for real-world OS coverage
- ~60% reduction in Linux build matrix jobs

+++ src/PD0.C
@@ -744,7 +744,7 @@ PD0::maxNumberOfCells(const std::string& fn)
       const size_t hdr(hdrLo | (hdrHi << 8));
       if (hdr == 0x0000) { // Fixed
         is.seekg(7, std::ios::cur); // Move to number of cells
-        const uint8_t nCells(is.get());
+        const uint8_t nCells(static_cast<uint8_t>(is.get()));
         maxnCells = (maxnCells > nCells) ? maxnCells : nCells;
         break; // Go to next block
       }
